Update CreateAPI to allow definitions with same ListenPath
CreateAPI checks whether any other API definitions listen to the same path, which is expected in case the base URL of the APIs is the same. You don't want overlapping endpoints. However, not all definitions have the same base URL (or Target URL), and if they don't, overlapping paths should be allowed. I think this check should be expanded to see if BOTH the target URL and listen path overlap with an already existing resource.