goreleaser does releases now, go1.15
After this is merged to master and is tested with a tag, the Buddy
pipelines that build and push packages should be deprecated.

This commit will push to tyk-sync-unstable if the triger is a
commit (not a tag). If it is a tag, packages are pushed to tyk-sync
and docker images are also published.